perf: eliminate N+1 and in-memory filtering across service/persistence layers
Replace per-item DAO lookups and full-table in-memory scans with batched and
DB-side queries, and extract one-shot legacy SQL out of startup schema.

Persistence (new DAO methods + queries):
- CarDao: findByBrandIdAndBodyTypeIdAndModel, findIdsByCriteria, and
  existsByBrandIdAndBodyTypeIdAndModelAndYearExcludingId push model lookup,
  candidate-id paging and duplicate detection into SQL/JPQL.
- Image DAOs: findByCarIdAndImageIdsWithData / findByReviewIdAndImageIdsWithData
  / findImagesByRequestId(AndImageIds)WithData fetch retained images in one
  query instead of one round-trip per id.
- ReviewDao.countByCarIdsSince, ReviewLike/ReviewReply countNew*PerReviewSince,
  CarFavoriteDao.findAllFavoriteCarIdsByUser and UserDao.findByIds enable the
  weekly digest to aggregate across all users in a handful of queries.
- CommunityJpaDao filters hidden comments in the query (was filtered in memory).

Services:
- RecommendationServiceImpl scores candidate ids and batch-loads only the
  top-N cars instead of paginating every matching car into memory.
- CarServiceImpl / CarRequestServiceImpl duplicate checks and image-payload
  building use the new batch/exists queries; getYearVariants ordered in SQL.
- WeeklyDigestServiceImpl builds review/favorite activity for all users from
  batched queries; moderator pending count uses countCarRequestsByStatus.
- AdminRequestServiceImpl/AdminController batch submitter user lookups.
- EmailService moderator-digest count widened int -> long.

Other:
- Extract legacy backfills and seed data from schema.sql into
  schema.deprecated.sql (not run at startup).
- CarReviewController uses request.isUserInRole("ADMIN").
- Remove redundant maven.compiler.release override in services/pom.xml.
- Tests updated/added for all new DAO and service behavior; qa.md route fix.

+++ services/src/main/java/ar/edu/itba/paw/services/EmailServiceImpl.java
@@ -320,7 +320,7 @@ public class EmailServiceImpl implements EmailService {
     @Override
     @Async("mailTaskExecutor")
     public void sendWeeklyModeratorDigest(final List<EmailRecipient> moderatorRecipients,
-                                          final int pendingRequestCount) {
+                                          final long pendingRequestCount) {
         if (moderatorRecipients == null || moderatorRecipients.isEmpty()) {
             return;
         }
@@ -481,20 +481,20 @@ public class EmailServiceImpl implements EmailService {
 
     // ── Weekly moderator digest ───────────────────────────────────────────────
 
-    private String buildModeratorDigestSubject(final int pendingCount, final Locale locale) {
+    private String buildModeratorDigestSubject(final long pendingCount, final Locale locale) {
         return pendingCount == 0
                 ? msg("email.moderatorDigest.subject.empty", locale, APP_NAME)
                 : msg("email.moderatorDigest.subject.pending", locale, APP_NAME, pendingCount);
     }
 
-    private String buildModeratorDigestPlainText(final int pendingCount, final Locale locale) {
+    private String buildModeratorDigestPlainText(final long pendingCount, final Locale locale) {
         if (pendingCount == 0) {
             return msg("email.moderatorDigest.plain.empty", locale, APP_NAME, appBaseUrl + "/admin");
         }
         return msg("email.moderatorDigest.plain.pending", locale, APP_NAME, pendingCount, appBaseUrl + "/admin");
     }
 
-    private String buildModeratorDigestHtml(final int pendingCount, final Locale locale) {
+    private String buildModeratorDigestHtml(final long pendingCount, final Locale locale) {
         final boolean hasPending = pendingCount > 0;
         final String preheader = hasPending
                 ? escapeHtml(msg("email.moderatorDigest.preheader.pending", locale, pendingCount))
@@ -513,7 +513,7 @@ public class EmailServiceImpl implements EmailService {
         );
     }
 
-    private String buildModeratorDigestStatusCard(final int pendingCount, final Locale locale) {
+    private String buildModeratorDigestStatusCard(final long pendingCount, final Locale locale) {
         if (pendingCount > 0) {
             return """
                     <div style="background:%s;border:1px solid %s;border-radius:18px;padding:28px 24px;margin-bottom:28px;text-align:center;">
